Break image serializers into smaller concern-grouped files
